news

+++ src/main/java/com/example/mysmallproject/service/HelperGenerics.java
@@ -3,6 +3,6 @@ import org.springframework.data.domain.Page;
 
 public interface HelperGenerics<T,C>{
     T save(C c);
-    T update(C c, int id);
-    void delete(int id);
+    T update(C c, long id);
+    void delete(long id);
 }
